# shared/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: DaisyReport API
  version: 1.0.0
  description: DaisyReport BI Platform REST API
  contact:
    name: Daisy Business Solutions
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://reports.daisybusiness.com
    description: Production
...
security:
  - BearerAuth: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token returned from /api/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key generated via /api/users/{id}/api-keys
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set after login

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Resource ID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number (1-based)
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 200
      description: Items per page
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search filter
    SortByParam:
      name: sortBy
      in: query
      schema:
        type: string
      description: Field name to sort by
    SortDirParam:
      name: sortDir
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: Sort direction

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 25
        totalItems:
          type: integer
          example: 142
        totalPages:
          type: integer
          example: 6
      required: [page, pageSize, totalItems, totalPages]

    # ---- Standard error ----
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Field 'email' is required"
      required: [status, code, message]

    # ---- Auth ----
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: "••••••••"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Refresh token for obtaining new access tokens
          example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        expiresIn:
          type: integer
          description: Token lifetime in seconds
          example: 3600
        user:
          $ref: "#/components/schemas/User"
      required: [token, refreshToken, expiresIn, user]
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8

    # ---- User ----
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@daisy.co.za
        displayName:
          type: string
          example: Admin User
        role:
          type: string
          enum: [admin, editor, viewer]
          example: admin
        orgUnitId:
          type: integer
          format: int64
          nullable: true
          example: 3
        enabled:
          type: boolean
          example: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, username, email, displayName, role, enabled, createdAt, updatedAt]
    UserCreate:
      type: object
      required: [username, email, displayName, role, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          example: jdoe
        email:
          type: string
          format: email
          example: jdoe@daisy.co.za
        displayName:
          type: string
          example: Jane Doe
        role:
          type: string
          enum: [admin, editor, viewer]
          example: editor
        password:
          type: string
          format: password
          minLength: 8
        orgUnitId:
          type: integer
          format: int64
          nullable: true
        enabled:
          type: boolean
          default: true
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer]
        orgUnitId:
          type: integer
          format: int64
          nullable: true
        enabled:
          type: boolean
    UserList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ---- API Key ----
...
  responses:
    BadRequest:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 400
            code: VALIDATION_ERROR
            message: "Field 'email' is required"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 401
            code: UNAUTHORIZED
            message: "Invalid or expired token"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 403
            code: FORBIDDEN
            message: "You do not have access to this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 404
            code: NOT_FOUND
            message: "Resource not found"
    Conflict:
      description: Conflict — duplicate or state violation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 409
            code: CONFLICT
            message: "A resource with that name already exists"
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 429
            code: RATE_LIMITED
            message: "Too many requests — retry after 60 seconds"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
...
paths:
...
  /api/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Authenticate with username and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Invalidate the current session and refresh token
      responses:
        "204":
          description: Logged out successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/session:
    get:
      operationId: getSession
      tags: [Auth]
      summary: Get current authenticated user info
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/change-password:
    post:
      operationId: changePassword
      tags: [Auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------------
  /api/users:
    get:
      operationId: listUsers
      tags: [Users]
      summary: List users (paginated, searchable)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/SearchParam"
        - $ref: "#/components/parameters/SortByParam"
        - $ref: "#/components/parameters/SortDirParam"
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, editor, viewer]
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      tags: [Users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/users/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      operationId: getUser
      tags: [Users]
      summary: Get user by ID
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateUser
      tags: [Users]
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Delete user
      responses:
        "204":
          description: User deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...